UserRepositoryをinterfaceに修正
