+++ app/src/main/java/com/example/learningandroidapp/di/UserRepositoryModule.kt
@@ -1,0 +1,15 @@
+package com.example.learningandroidapp.di
+
+import com.example.learningandroidapp.repository.UserRepository
+import com.example.learningandroidapp.repository.UserRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class UserRepositoryModule {
+    @Binds
+    abstract fun bindUserRepository(impl: UserRepositoryImpl): UserRepository
+}
