+++ app/src/main/java/com/example/learningandroidapp/repository/UserRepository.kt
@@ -1,23 +1,8 @@
 package com.example.learningandroidapp.repository
 
-import com.example.learningandroidapp.api.GitHubApiService
 import com.example.learningandroidapp.models.User
-import javax.inject.Inject
-import javax.inject.Singleton
 
 
-@Singleton
-class UserRepository @Inject constructor(
-    private val gitHubApi: GitHubApiService
-) {
-    suspend fun getUserList(text: String): List<User> {
-        val response = gitHubApi.getUsers(text)
-        val userList = response.items.map {
-            User(
-                userName = it.login,
-                avatarUrl = it.avatarUrl
-            )
-        }
-        return userList
-    }
+interface UserRepository {
+    suspend fun getUserList(text: String): List<User>
 }
